refactor: consolidate and decompose the codebase for the 0.5.0 release
- Narrow seven over-public items to `pub(super)` or private, dropping `ParseRuleIdError`'s public field
- Gate `From<&str> for RuleId` behind `cfg(test)`, a constructor only tests reach
- Lift the line-fit check into `Source::column_overflows`, shared across three alignment rules
- Collapse duplicated logic into `CleanReport::record`, `statement_annotations`, and an itertools `join`
- Decompose `binding.rs` and `aligner/groups.rs` into directory modules of finer sibling files
- Split `config/mod.rs`'s test module into `parse_tests` and `load_tests` siblings
- Rename four colliding fixture cases to tree-wide-unique names
