+++ src/diagnostics/mod.rs
@@ -50,14 +50,12 @@ pub(crate) struct EmitterSummary {
 
 /// Flattens every run into its `(file, diagnostic)` pairs in
 /// file-major order, the traversal each emitter walks.
-pub(crate) fn diagnostics<'a>(
-    runs: &'a [Run<'a>],
-) -> impl Iterator<Item = (&'a SourceFile, &'a Diagnostic)> {
+fn diagnostics<'a>(runs: &'a [Run<'a>]) -> impl Iterator<Item = (&'a SourceFile, &'a Diagnostic)> {
     runs.iter()
         .flat_map(|(file, ds)| ds.iter().map(move |d| (*file, d)))
 }
 
-pub(crate) fn line_columns(file: &SourceFile, range: TextRange) -> (LineColumn, LineColumn) {
+fn line_columns(file: &SourceFile, range: TextRange) -> (LineColumn, LineColumn) {
     let code = file.to_source_code();
     (
         code.line_column(range.start()),
@@ -65,7 +63,7 @@ pub(crate) fn line_columns(file: &SourceFile, range: TextRange) -> (LineColumn, 
     )
 }
 
-pub(crate) fn write_json_line<T: Serialize>(writer: &mut dyn Write, value: &T) -> io::Result<()> {
+fn write_json_line<T: Serialize>(writer: &mut dyn Write, value: &T) -> io::Result<()> {
     serde_json::to_writer(&mut *writer, value).map_err(io::Error::other)?;
     writer.write_all(b"\n")
 }
