+++ src/rules/align_match_case.rs
@@ -107,8 +107,11 @@ impl Visitor<'_> {
             .slice(TextRange::new(case.start(), pre_colon_end))
             .width();
         let body_width = self.walker.source.slice(body_first.range()).width();
-        self.walker.source.column_of(case.start()) + lhs_width + 3 + body_width
-            > self.code_line_length
+        self.walker.source.column_overflows(
+            case.start(),
+            lhs_width + 3 + body_width,
+            self.code_line_length,
+        )
     }
 
     /// Emits collapse-and-align edits for one match by walking each
