+++ tests/fixtures/align_comparisons/own_line_comment_breaks_operator_run/input.py
