+++ tests/fixtures/alphabetize/parenthesized_value_keeps_parens_on_reorder/input.py
