+++ tests/fixtures/docstring_wrap/within_budget_docstring_untouched/input.py
