+++ tests/fixtures/docstring_wrap/within_budget_docstring_untouched/meta.toml
