+++ tests/fixtures/signature_layout/parameter_comment_pins_shape/input.py
